changed to tsx
