--- client/editor-ai/app/homepage/Layout.jsx
@@ -1,22 +1,0 @@
-import { Outlet, Link } from "react-router-dom";
-
-const Layout = () => {
-  return (
-    <>
-      <nav>
-        <ul>
-          <li>
-            <Link to="/homepage">Home</Link>
-          </li>
-          <li>
-            <Link to="/texteditor">Text Editor</Link>
-          </li>
-        </ul>
-      </nav>
-
-      <Outlet />
-    </>
-  )
-};
-
-export default Layout;
