+++ client/editor-ai/app/homepage/ProjectItem.tsx
