+++ client/editor-ai/app/texteditor/SuggestionBox.tsx
